Update README.md
We are looking for people to join our DAO team forging a new path for Decentralized Autonomous Organizations, bringing Real World, Sustainable Patented Green Energy Technologies that turn wastes and garbage into Highly Valuable Electricity and Abundant Food and Water Commodities, without producing any carbon emissions or pollutions of any kind. In Fact, the Air & Water in and around the Power plants is cleaner coming out, than it was going in!

Join us in this noble and epic adventure to allow Everyone, Everywhere in the world the opportunity to own and participate in the Governance, production and profits of this Up and Coming Business!